### 2021-06-14 #3367 @whedon
[PRE REVIEW]: Kinematics-vis: A Visualization Tool for the Mathematics of Human Motion
**Submitting author:** @heath-henninger (<a href="http://orcid.org/0000-0002-3635-1289">Heath Henninger</a>)
**Repository:** <a href="https://github.com/klevis-a/kinematics-vis.git" target ="_blank">https://github.com/klevis-a/kinematics-vis.git</a>
**Version:** v.1.0.0
