### 2021-06-14 #3367 @heath-henninger
From: whedon ***@***.***>
Sent: Monday, June 14, 2021 6:27 AM
Cc: Heath Byron Henninger ***@***.***>; Mention ***@***.***>
Subject: [openjournals/joss-reviews] [PRE REVIEW]: Kinematics-vis: A Visualization Tool for the Mathematics of Human Motion (#3367)

Submitting author: @heath-henninger<https://github.com/heath-henninger> (Heath Henninger<http://orcid.org/0000-0002-3635-1289>)
Repository: https://github.com/klevis-a/kinematics-vis.git
Version: v.1.0.0
Editor: Pending
Reviewer: Pending
Managing EiC: Daniel S. Katz

⚠️ JOSS reduced service mode ⚠️

Due to the challenges of the COVID-19 pandemic, JOSS is currently operating in a "reduced service mode". You can read more about what that means in our blog post<https://blog.joss.theoj.org/2020/05/reopening-joss>.

Status

[status]<https://joss.theoj.org/papers/ef7f95fc8216dd7c6d37a6679374b175>

Status badge code:

Author instructions

Thanks for submitting your paper to JOSS @heath-henninger<https://github.com/heath-henninger>. Currently, there isn't an JOSS editor assigned to your paper.

@heath-henninger<https://github.com/heath-henninger> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this list of people<https://bit.ly/joss-reviewers> have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Editor instructions

The JOSS submission bot @whedon<https://github.com/whedon> is here to help you find and assign reviewers and start the main review. To find out what @whedon<https://github.com/whedon> can do for you type:
