### 2021-08-01 #3490 @heath-henninger
A valid comment Nicholas.  As the PI on the grant that collected these kinematic data and stimulated the generation of the visualization and analysis application, I typically submit the manuscripts for peer-review.  In this case Klevis, a recent PhD graduate from my lab, was the lead investigator who created the code in support of his dissertation research.

This is definitely a case of our naiveté about the JOSS process versus the typical peer reviewed journal.  I sense this requires us to switch the submitting author, but will this require the whole review process to begin again?  Or can we shortcut the process with a simple assignment change?

Any information is helpful.

Heath

From: Nicholas Nadeau, Ph.D., P.Eng. ***@***.***
Sent: Sunday, August 1, 2021 8:52 AM
Cc: Heath Byron Henninger ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Kinematics-vis: A Visualization Tool for the Mathematics of Human Motion (#3490)

@adi3<https://github.com/adi3> , before getting started, I think there may be an issue with the "submitting author":

  *   the review has @heath-henninger<https://github.com/heath-henninger> as the submitting author, but it's an inactive account with no contributions to the repo
  *   @klevis-a<https://github.com/klevis-a> is the sole contributor to the repo, but isn't mentioned anywhere here
