### 2021-08-01 klevis-a/kinematics-vis#2 @nnadeau
JOSS: Usage instructions
- The README currently only contains installation instructions
- The README (and potentially the paper) should also include a `Usage` section that guides the user into how to use the software, especially considering that:
  - "Human motion analysis is a complex topic. Biomechanists, clinicians, and trainees typically accept the results of research papers without having a consistent way to validate these results"
  - "It has previously been shown that incorrect usage of Euler/Cardan sequences can significantly misrepresent human motion"
  - "biomechanics researchers and clinicians with a limited mathematical background"
- cc https://github.com/openjournals/joss-reviews/issues/3490
