### 2021-08-01 klevis-a/kinematics-vis#3 @nnadeau
JOSS: Explain motion datasets
- Given #2 and that this repo "provides a means for researchers to make their motion capture datasets visually explorable"
- What datasets? What requirements or limitations?
- cc https://github.com/openjournals/joss-reviews/issues/3490
