### 2021-08-01 klevis-a/kinematics-vis#4 @nnadeau
JOSS: Use standard packaging and installation approaches
- Using standard packaging and installation approaches helps foster external collaboration and reduces issues/errors, especially when robust tools with their own checks are leveraged (e.g., npm, yarn, make, pip)
- This is especially important for security and SecOps, which is very strict in clinical/medical environments
- The current installation approach has the user running untrusted scripts, downloading artifacts from external servers, modifying files/folders/paths, unzipping archives, and running powershell subprocesses
- Many of the above steps would be flagged in a (properly) secure IT environment (e.g., powershell is often blocked on most PCs in well-maintained IT environments)
- Moverover, there is code duplication between windows vs linux installation scripts (e.g., forward vs backslashes)
- If standard installation or packaging tools were used, duplications and security issues would be resolved, as these tools are platform agnostic
- Given that `downloadDeps.sh` appears to focus on js installation, I'd recommend a js-based package manager (e.g., npm or yarn)
- cc https://github.com/openjournals/joss-reviews/issues/3490
