### 2021-08-01 klevis-a/kinematics-vis#5 @nnadeau
JOSS: Add `Contributing` section to README
- [Standard README sections](https://guides.github.com/features/wikis/#Formatting-a-readme)
- cc https://github.com/openjournals/joss-reviews/issues/3490
