### 2021-08-01 klevis-a/kinematics-vis#6 @nnadeau
JOSS: Alternative software and approaches?
- The paper only mentions that researchers have typically used 2D graphs
- Is there no other visualization software available?
- Has there really been no work done before this repo?
- Citations will be needed to validate the point
- Pros, cons, and limitations of this repo vs other approaches should also be discussed
- cc https://github.com/openjournals/joss-reviews/issues/3490
