### 2021-08-01 klevis-a/kinematics-vis#7 @nnadeau
JOSS: Mathematics section
- While referring the reader to relevant sources is a good approach to keeping the paper concise, the section needs to be filled with at least an overview and a discussion of why each of the sources listed are relevant
- cc https://github.com/openjournals/joss-reviews/issues/3490
