### 2021-08-01 klevis-a/kinematics-vis#8 @nnadeau
JOSS: Novelty of repo
- Visualization and providing tools is great, but **why** should a researcher use this particular repo?
- "biomechanics researchers and clinicians [may have] a limited mathematical background", but we're assuming that they're skilled enough to use github and run shell scripts
- Why wouldn't they program their own simple visualizer? `three.js` is super popular and supported, why not create a simple visualizer for just your exact need?
- Why would a researcher depend on an external repo? Each external tool or dependency that's used carries its own risk; is this repo worth the risk?
- Has this repo been used in published research to prove its novelty and robustness?
- cc https://github.com/openjournals/joss-reviews/issues/3490
