### 2021-08-01 klevis-a/kinematics-vis#9 @nnadeau
JOSS: Add  Code of conduct 
- https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project
- cc https://github.com/openjournals/joss-reviews/issues/3490
