### 2021-08-01 klevis-a/kinematics-vis#11 @nnadeau
JOSS: Add Pull request template 
- If this repo will be used by "biomechanics researchers and clinicians with a limited mathematical background", we'll need to guide them properly
- https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository
- cc https://github.com/openjournals/joss-reviews/issues/3490
