### 2021-08-01 klevis-a/kinematics-vis#12 @nnadeau
JOSS: Add tests and CI/CD
- Tests can be used for both CI/CD and for new users to ensure they properly installed the repo
- Tests should be based on example datasets that should also be provided
- Tests should be run for all PRs to ensure quality
- If this repo will be used by "biomechanics researchers and clinicians with a limited mathematical background", we'll need to guide them properly
- cc https://github.com/openjournals/joss-reviews/issues/3490
