### 2021-08-01 klevis-a/kinematics-vis#13 @nnadeau
JOSS: Add versioning
- Versioning is especially important in research, where exact versions of software/tools should always be cited
- This repo currently has no releases or versioning
- SemVer or CalVer are recommended
- cc https://github.com/openjournals/joss-reviews/issues/3490
