### 2021-08-01 #3490 @nnadeau
@adi3 @klevis-a, i did a first pass of the repo from a software architecture point of view, see the issues linked above
