### 2021-08-01 #3490 @adi3
@nnadeau thank you for bringing up the point about the submitting author. @heath-henninger appreciate the additional clarity you provided. We do expect the submitting author to have made substantial contributions to the repository, but I will make an exception to continue with this review for two reasons:

1. @klevis-a, who made all (or at least most of) the commits to the repository, is still listed as the primary author of the paper, and
2. @heath-henninger did have a role in the authorship and contributed to the work that went in the paper; it seems now he is simply handling the logistics of the publication process.

Separately -
@klevis-a please see the list of issues opened above and let @nnadeau know when you have them ready for review.

Thanks everyone for your contribution so far!
