### 2021-08-02 #3490 @klevis-a
@nnadeau thank you for taking time to review our submission. @nnadeau @adi3, this software was written to help me visually understand how we quantify human motion. It provided much of the insight that helped us publish "Beyond Euler/Cardan analysis: True glenohumeral axial rotation during arm elevation and rotation". Furthermore, other biomechanics researchers have found it helpful for understanding and teaching Euler/Cardan angles. I have already finished my graduate studies as of May of this year and moved on. And, I don't have the time to address the issues regarding standard packaging, unit tests, and continuous integration. Writing good unit tests can take as long (if not longer) than writing the actual software. I do believe that these topics are important, but unfortunately my time is limited.

If you would still consider the software for publication without me having to address the unit tests, continuous integration, and standard packaging then I can start addressing the other issues. If not, then I understand the choice. The journal (rightfully so) may need to uphold certain standards, but I won't be able to fulfill them.
