### 2021-08-03 #3490 @adi3
@klevis-a thank your for the clear response.

@heath-henninger do you have alternate resources to fix the editorial issues raised in this review, or perhaps you could do them yourself?
