### 2021-08-03 #3490 @heath-henninger
Unfortunately not. Klevis was a bit of a unicorn in my research team with his extensive background in code and software development.  I hired him specifically because he could do great things that my team was otherwise unprepared to accomplish. Believe that we will feel his absence now that he has graduated.

And this includes me. I am an experimental biomechanist by trade…so my coding experience ends at Matlab and some FORTRAN in the early 90s :-)
