### 2021-08-03 #3490 @adi3
@klevis-a I've discussed this with the editorial team and this is how we can proceed -

- You do not need to go with standard packaging, as long as installation instructions are clearly documented in the Readme.
- In the absence of automated tests (and CI/CD), we would at least need documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior).

I believe you'll be able to work on the other issues as is. Let us know once you've fixed all the tickets opened by @nnadeau.
