### 2021-08-03 #3490 @klevis-a
@adi3 thank you for being flexible! Yes, I can provide manual steps that can be followed to objectively check the expected functionality of the software. And I'll start working on the other tickets.
