### 2021-08-03 klevis-a/kinematics-vis#2 @klevis-a
I had included usage instructions in the UI. See https://shouldervis.chpc.utah.edu/kinevis/main.html as an example. I can also include them in the README if you think it is necessary. Please let me know.
