### 2021-08-03 klevis-a/kinematics-vis#4 @klevis-a
Closing this based on our discussion in openjournals/joss-reviews#3490.
