### 2021-08-03 klevis-a/kinematics-vis#2 @engnadeau
ah, i completely missed the `?` at the top

![image](https://user-images.githubusercontent.com/6395915/128056548-8082049e-4a09-48e2-b9dd-2bbd46c0923e.png)

yeah, it'd be great if the README at least directed the user where to look
