### 2021-08-03 klevis-a/kinematics-vis#2 @klevis-a
Yep, exactly. It's not immediately obvious - I was trying to conserve screen real estate. Sure, I can add that to the README.
