### 2021-08-03 klevis-a/kinematics-vis#10 @klevis-a
Addressed in ae753f020cfc80696ee74e28224fb259153975ef.
