### 2021-08-03 6dc0d8a
Addressed issue #6: alternative software and approaches
