+++ joos/paper.md
@@ -41,6 +41,10 @@ Kinematic analysis studies characterize human motion in healthy, pathologic, and
 
 ‘Kinematics-vis’ is a JavaScript web application that visualizes different methods of quantifying and decomposing human motion including, but not limited to, Euler/Cardan angles \autoref{fig:fig1}. The software enables researchers to visualize different kinematic analysis methods while outputting the 2D time-series graphs that are typically presented in peer-reviewed manuscripts [@Ludewig:2009]. Presently, ‘kinematics-vis’ is specialized to study shoulder kinematic analysis. However, it is easily expandable to accommodate the decomposition sequences of other joints, including subject-specific 3D anatomic models. ‘Kinematics-vis’ is designed to work from a central motion capture dataset, but also enables the visualization of singular motion files. JavaScript and WebGL, the technologies which underlie ‘kinematics-vis’, are ubiquitous and a standard in modern web browsers – which facilitates collaboration between researchers and clinicians.
 
+A variety of commercial musculoskeletal modeling software are utilized in biomechanics, including [AnyBody](https://www.anybodytech.com/) and [Visual3D](https://www.c-motion.com/#visual3d). Computational kinematics packages (e.g. rotation matrices, trajectory generation, quaternions) are abundant and are typically provided as part of scientific computing environments (e.g. [SciPy Spatial Transform package](https://docs.scipy.org/doc/scipy/reference/spatial.transform.html)) - but, visualization software is scarce. The open-source musculoskeletal modeling software [OpenSim](https://opensim.stanford.edu/work/index.html) does provide the ability to visualize the movement of bones and joints. However, what is missing from all the aforementioned software is the ability to visualize how decomposition methods (particularly Euler/Cardan angles) quantify human motion. Because 12 Euler/Cardan angle sequences exist, the choice of such a sequence can have a meaningful effect on the kinematic plots presented in biomechanics studies to researchers and clinicians [@Aliaj:2021; @Baker:2011]. Notably, [@Baker:2011] visually demonstrates how the measured joint angles are affected by the chosen Euler/Cardan angle sequence and argues that "only when we can adequately visualise these angles that we can engage in meaningful consideration of which orientation this is." Although [@Baker:2011] represents the first attempt at visualizing these angles in biomechanics literature, the presented analysis is static. It consists of a series of screenshots in a journal article.
+
+The solution presented herein enables biomechanics researchers to visualize the 3D angles that Euler/Cardan decompositions quantify, and to associate these 3D angles with typical 2D plots. The interactive nature of the software also enables exploration that is not possible within the confines of a journal article page. The presented software can be especially useful to novice biomechanics researchers (e.g. undergraduate and graduate students) for learning kinematic analysis. The downside of the presented software is that it offers only kinematic analysis within the musculoskeletal modeling and analysis domain. Perhaps a better approach is to include this software as a plug-in to [OpenSim](https://opensim.stanford.edu/work/index.html). The downside of the plug-in approach is that researchers that perform kinematic analysis may not perform subsequent modeling. Regardless, it would be a good idea to consider whether future iterations of this software should be developed as a plug-in to OpenSim.
+
 Human motion analysis is a complex topic. Biomechanists, clinicians, and trainees typically accept the results of research papers without having a consistent way to validate these results. ‘Kinematics-vis’ enables discussion amongst all parties interested in human motion analysis and obviates the need for a significant mathematical background. ‘Kinematics-vis’ provides a means for researchers to make their motion capture datasets visually explorable.  This provides insight which is missing in the 2D graphs accompanying peer-reviewed publications. This software can be used as an adjunct to published manuscripts to allow readers to analyze a motion of interest both visually and quantitatively. 
 
 # Mathematics
