### 2021-08-03 klevis-a/kinematics-vis#8 @klevis-a
@nnadeau see https://github.com/klevis-a/kinematics-vis/commit/6dc0d8a522589cb835b86a5b44789657b21eea5e. I'm happy to create Pull Requests for these, if that's easier for you.
