### 2021-08-03 klevis-a/kinematics-vis#8 @klevis-a
@nnadeau sorry, the aforementioned commit doesn't address this issue. It addresses another one, please ignore.
