### 2021-08-03 klevis-a/kinematics-vis#6 @klevis-a
@nnadeau https://github.com/klevis-a/kinematics-vis/commit/6dc0d8a522589cb835b86a5b44789657b21eea5e should address these questions.
