### 2021-08-03 klevis-a/kinematics-vis#8 @klevis-a
Hmm, now I wonder if the aforementioned commit does address these concerns as well. Please let me know. These are my thoughts regarding your comments.

1. The gap between creating 3D visualization/animations using three.js and going to github and running a shell script is very wide, in my opinion. The former requires knowledge of JavaScript, HTML, CSS, three.js, and 3D rotations (which are notoriously confusing). The latter is under the knowledge domain of most researchers. However, I do agree that modifying the code in order create custom functionality will require knowledge of JavaScript, three.js, etc. However, at that time this repo provides a decent starting base for researchers.

2. This repo does not create a code dependency (i.e. it is designed to be utilized as a standalone utility). The only downside that I see is that dependencies need to be downloaded via a shell script. However, if someone wants to circumvent this - it's easy enough to download the dependencies manually. They total under 10 files.
