### 2021-08-03 klevis-a/kinematics-vis#11 @klevis-a
Done in https://github.com/klevis-a/kinematics-vis/commit/935619b12d7e1126cfe8b78d4bea73bd164d2d76.
