### 2021-08-03 a643961
Created CODE_OF_CONDUCT.MD
