### 2021-08-03 klevis-a/kinematics-vis#9 @klevis-a
Addressed in https://github.com/klevis-a/kinematics-vis/commit/a6439613fc8712b44c05b97a2662c92804b9efba
