### 2021-08-03 14b6b21
Adding name of project and Usage section to README.
