### 2021-08-03 klevis-a/kinematics-vis#2 @klevis-a
Added a couple of sentences directing users where to find usage instructions in https://github.com/klevis-a/kinematics-vis/commit/14b6b210d3c0292f3b27e713cd33c54ae400344c.
