### 2021-08-03 e0dceb4
The activity names and their order of appearance in the UI are now explicitly specified via the database JSON file.
