### 2021-08-03 d8e0aee
The capture frequency for a particular trajectory is now set explicitly and not assumed.
