### 2021-08-03 ae4a033
Updated README to show how to select a trial to analyze from the sample dataset and to point to INPUT_FILES.md, which explains the file formats utilized in this web app.
