### 2021-08-03 klevis-a/kinematics-vis#3 @klevis-a
Dataset format is now explained in INPUT_FILES.md, see https://github.com/klevis-a/kinematics-vis/commit/ae4a033f2f08b2e4e6df986bb4d18add7746cffc.
