### 2021-08-04 2a9cde8
Added CONTRIBUTING document and updated README.
