+++ CONTRIBUTING.md
@@ -1,0 +1,38 @@
+# Contributing to Kinematics-Vis
+We want to make contributing to this project as easy and transparent as possible, whether it's:
+
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+- Becoming a maintainer
+
+## We Develop with Github
+We use github to host code, to track issues and feature requests, as well as accept pull requests.
+
+## We Use [Github Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
+Pull requests are the best way to propose changes to the codebase. We actively welcome your pull requests.
+
+## Any contributions you make will be under the MIT Software License
+In short, when you submit code changes, your submissions are understood to be under the same [MIT License](http://choosealicense.com/licenses/mit/) that covers the project.
+
+## Report bugs using Github's [issues](https://github.com/klevis-a/kinematics-vis/issues)
+We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/klevis-a/kinematics-vis/issues/new?assignees=&labels=&template=bug_report.md&title=); it's that easy!
+
+## Write bug reports with detail, background, and sample code (if possible)
+
+**Great Bug Reports** tend to have:
+
+- A quick summary and/or background
+- Steps to reproduce
+  - Be specific!
+  - Give sample code if you can.
+- What you expected would happen
+- What actually happens
+- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+
+## License
+By contributing, you agree that your contributions will be licensed under its MIT License.
+
+## References
+This document was adapted from the following open-source [contribution guidelines](https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62).
