### 2021-08-04 klevis-a/kinematics-vis#5 @klevis-a
This is now complete: https://github.com/klevis-a/kinematics-vis/commit/2a9cde872fabe0c2dad87504ebef12483c4082f8
