### 2021-08-04 36b7553
Addressed issue #7: Mathematics section is now multiple paragraphs that explain how the list of references is useful for understanding the mathematics behind the web app.
