### 2021-08-04 klevis-a/kinematics-vis#7 @klevis-a
Addressed in https://github.com/klevis-a/kinematics-vis/commit/36b7553d6086dc6f7d01d2cc24d1452f0ecb879c.
