### 2021-08-05 cdd7dfb
Adding basic manual test cases.
