+++ MANUAL_TESTING.md
@@ -1,0 +1,206 @@
+Before proceeding with the test cases below, follow the instructions in the [README](README.md) to install the software and download the sample dataset.
+
+### Test Cases
+
+#### 1. Open a motion trial
+
+**Steps**
+
+1. Click the folder button (top of upper left hand quadrant) to open the trial selection dialog.
+2. Select subject `O45_001_F_47_R` and `Scapular Plane Abduction`.
+3. Click Analyze.
+
+**Expected Outcome**
+
+The UI should briefly flash `Loading...` then the trial analysis should be loaded.
+
+* The upper left pane should have the `Humerus ISB: yx'y''` view.
+* The upper right pane should have the `Scapula ISB: yx'z''` view.
+* The lower left pane should have the `Preview` view.
+* The lower right panel should contain plots and the default plot selected should be `Humerus ISB: yx'y''`.
+
+#### 2. Orbital Controls
+
+In all non-plot views the following controls should work:
+
+1. Left-button dragging rotates the view.
+2. Right-button dragging pans the view.
+3. Middle-button scrolling zooms in and out.
+
+#### 3. Motion capture frame preview (slider bar)
+
+**Steps**
+
+Drag the button of the slider bar (middle top of UI) to the right.
+
+**Expected Outcome**
+
+* The frame number indicator to the right of the slider bar should increase.
+* All non-plot views should be updated. The slightly transparent scapula and humerus should be elevating.
+
+#### 4. Analyzing a particular motion capture frame (slider bar)
+
+**Steps**
+
+Press the downward facing arrow to the left of the slider bar (once a motion capture frame has been selected via the slider bar).
+
+**Expected Outcome**
+
+* All views that start with Humerus or Scapula should be updated. The scapula view does not have a visual indication that updating completed successfully (subsequent test cases can validate that the update did complete successfully, however). The current longitude and latitude lines (indicated by yellow) on the Humerus views will update.
+
+#### 5. Motion capture frame preview (plots)
+
+**Steps**
+
+Hover the mouse button over any plot curve. Trace the curve with the mouse pointer towards the right of the screen.
+
+**Expected Outcome**
+
+* The frame number indicator to the right of the slider bar should increase.
+* All non-plot views should be updated. The slightly transparent scapula and humerus should be elevating.
+
+#### 6. Analyzing a particular motion capture frame (plots)
+
+**Steps**
+
+Once a particular motion capture frame of interest has been determined (by hovering over the plot curves), click the left mouse button.
+
+**Expected Outcome**
+
+* All views that start with Humerus or Scapula should be updated. The scapula view does not have a visual indication that updating completed successfully (subsequent test cases can validate that the update did complete successfully, however). The current longitude and latitude lines (indicated by yellow) on the Humerus views will update.
+
+#### 7. Switching the displayed view in a quadrant
+
+**Steps**
+
+Apart from the lower right quadrant, the views for all other quadrants can be switched via their drop-down selector. For each of the other 3 quadrants, make sure that each of the following views can be selected:
+
+1. `Humerus ISB: yx'y''`
+2. `Humerus Phadke: xz'y''`
+3. `Humerus Swing Twist`
+4. `Humerus Simultaneous`
+5. `Preview`
+6. `Scapula ISB: yx'z''`
+
+**Expected Outcome**
+
+All of the quadrants apart form the lower right one should update once the view that it displays is updated via the drop-down box.
+
+#### 7. Correct number of steps for each view
+
+**Steps**
+
+This test case assures that the correct number of steps are associated with each view. The number of steps can be seen on the lower left corner of each view.
+
+**Expected Outcome**
+
[…]
