### 2021-08-05 klevis-a/kinematics-vis#12 @klevis-a
Per our discussion in openjournals/joss-reviews/#3490 added manual test cases in https://github.com/klevis-a/kinematics-vis/commit/cdd7dfb7187f948c60f479761c9c99f0692e5543.
