### 2021-08-05 klevis-a/kinematics-vis#13 @klevis-a
Released version 0.1.0 of software.
