### 2021-08-05 #3490 @klevis-a
@nnadeau @adi3 I have now addressed the currently open issues. I have left them open since I thought they would be easier to track this way. @nnadeau feel free to close them as you review the changes. Otherwise, I can close them - whichever makes this process easier.
