### 2021-08-09 59dc3fc
This fixes the issue where the plotly plots were not being properly resized. This happened because the plotly plots only get resized on window size changes. So this fix dispatches a resize event when double click happens.
