### 2021-08-10 klevis-a/kinematics-vis#13 @engnadeau
- closing issue
- fyi any further changes to the codebase for JOSS could potentially bump the version
