### 2021-08-10 klevis-a/kinematics-vis#12 @engnadeau
@klevis-a

- would it be possible to make these changes in PRs for easier review and traditional git flow?
- where are sample files from?
