### 2021-08-10 klevis-a/kinematics-vis#8 @engnadeau
Thanks for adding the section. Has this repo been used in published research to demonstrate its novelty?
