### 2021-08-10 klevis-a/kinematics-vis#2 @engnadeau
could you also provide a screenshot similar to mine in the README to visually show people where the help menu is
