### 2021-08-10 klevis-a/kinematics-vis#8 @klevis-a
It has been utilized in our publication titled "Beyond Euler/Cardan Analysis: True Axial Rotation...". However, at the time I could not cite the repo because it was under review at the technology transfer office (PIVOT). We have received positive feedback from it at a recent conference and from other colleagues. I think it's very useful from a learning perspective. As far as I can tell, there's no other tool in biomechanics that visualizes decompositions (plenty of them will visualize the motion of the bones).

I could understand how you would think that a tool like this should exist already, but the unfortunately reality is that it doesn't. I would not have created this utility otherwise.
