### 2021-08-10 klevis-a/kinematics-vis#12 @klevis-a
@nnadeau

* Are you referring to all the issues that you opened (and are now closed)? If so, yes I can create PRs for these.
* The sample files (hosted at the University of Utah Center for High Performance Computing) are downloaded via the downloadData scripts. The dataset was collected in our lab. It was published alongside "Beyond Euler/Cardan angles: True Axial Rotation..." in [Zenodo](https://zenodo.org/record/4536684). However, the dataset downloaded via the downloadData scripts is not identical to the one on Zenodo because it conforms to the format specification of the webapp.
