### 2021-08-10 f296fc1
Adding image showing where the Help button is located.
