### 2021-08-10 klevis-a/kinematics-vis#2 @klevis-a
Added screenshot to README in https://github.com/klevis-a/kinematics-vis/commit/f296fc1fcba70285e3a03fe3a9d37874e364b644.
