### 2021-08-11 klevis-a/kinematics-vis#12 @engnadeau
yeah, i was referring to all the commits made. Ideally, the commits aren't made directly to `master`, but rather a PR is created per issue (or batch of issues). I.e., PRs should close issues, not direct commits.
