### 2021-08-11 klevis-a/kinematics-vis#12 @klevis-a
@nnadeau I initially thought that PRs could also be made retroactively between any commits - but that's not the case. It seems that PRs can only be made between two branches. I'll keep that in mind moving forward. I'm not sure what to do about the prior commits. There might be a way to make them into PRs but it won't be straightforward.
