### 2021-09-13 klevis-a/kinematics-vis#14 @destogl
JOSS: Can not load test data
openjournals/joss-reviews#3490

**Describe the bug**
When trying to open any test data, "File not found" error happens.
Using `master` branch.

Hint about error: It appears that the selected data in the web browser are not filled correctly when trying to access them.

**To Reproduce**
Steps to reproduce the behavior:
1. Install as per README, download dependencies and data using scripts.
2. Start python server as per README
3. open localhost:8000/man.html

**Expected behavior**
Test data should be shown in the view, instead "Loading..." is forever on the scrren.

 - OS: Ubuntu 20.04
 - Browser Firefox 91.0.2; Chromium 93.0.4577.63

**Additional context**
Terminal output:
```
127.0.0.1 - - [13/Sep/2021 11:45:53] code 404, message File not found
127.0.0.1 - - [13/Sep/2021 11:45:53] "GET /csv/undefined HTTP/1.1" 404 -
127.0.0.1 - - [13/Sep/2021 11:45:53] code 404, message File not found
127.0.0.1 - - [13/Sep/2021 11:45:53] "GET /csv/undefined HTTP/1.1" 404 -
```

Under `csv` folder, I can see the data.
