### 2021-09-13 klevis-a/kinematics-vis#15 @destogl
JOSS: Consolidate default branch
As per #14 it seems that `master` branch is not working, but `joss` does. Before finalizing the review, I think branches should be consolidated and the default branch a user is checking out has to function properly. There are multiple options to resolve this:

- Make `joss` branch default (and maybe remove `master` branch)
- Merge `joss` into master.

cc. openjournals/joss-reviews#3490
