### 2021-09-13 #3490 @destogl
@adi3 Thanks for your patience. The holiday season behind brought hard-time to focus on this review. Next time I have to plan better and let you know.

@klevis-a, @heath-henninger: Here are my overall comments on software:

- The live code demo was very useful. Thanks for creating it. This is especially useful for target users.
- Regarding "Substantial scholarly effort": my main area of research and interest is robotics, although in my research I was "touching" kinematics of humans, I don't have in-depth knowledge of the field. Nevertheless, the kinematics challenged the software is addressing is very well known and discussed in robotics, and there we manged to find a suitable standard for data-representation. Therefore, I clearly see contribution of this software and regarding discussion in klevis-a/kinematics-vis#8 I am convicted in added values for the biomechanics research community.

Please also consider a newly opened issues.
