### 2021-09-13 fcc5ea5
Removing ActivityFriendlyNames - this is no longer needed since these are specified in the JSON file.
