### 2021-09-13 klevis-a/kinematics-vis#14 @klevis-a
@destogl sorry about that. During the review process I made changes to the master branch in order to simplify data specification. I updated the live link dataset, but forgot to update the test data dataset. Really they are the same thing, but the test data is a zipped version for easy downloading. I simply forgot to update the zip file. This update is done now - no code changes needed. Please test it out. You can simply delete your current csv directory and run ./downloadData.sh again.
