### 2021-09-13 klevis-a/kinematics-vis#15 @klevis-a
@destogl the issue that you ran into with #14 was due to me not having updated the test dataset. All code changes that I have made during the review have been to master - that is the branch that is the most up to date. The joss branch is simply there to contain the paper. And I believe this is fine, according to the JoOS instructions.
