### 2021-09-13 #3490 @klevis-a
@destogl thank you for chiming in on the contribution of the software - I appreciate it. We also think that it adds value to the biomechanics community, but of course, we are a bit biased.

The software has been especially useful in convincing other biomechanics researchers that Euler angles cannot be added or subtracted. My first Ph.D. project was in robotics so this was almost a given to me. But the practice of subtracting Euler angles is still widely practiced in the biomechanics community.
