### 2021-09-14 bbf0560
Added references for software packages mentioned in paper. Also added references for Euler/Cardan angles.
