+++ joos/paper.bib
@@ -107,3 +107,65 @@
    doi = {10.1007/s11517-018-1903-3},
    url = {https://pubmed.ncbi.nlm.nih.gov/30367391/}
 }
+
+@article{Delp:2007,
+   author = {Delp, S.~L. and Anderson, F.~C. and Arnold, A.~S. and Loan, P. and Habib, A. and John, C.~T. and Guendelman, E. and Thelen, D.~G.},
+   title = {OpenSim: Open-Source Software to Create and Analyze Dynamic Simulations of Movement},
+   journal = {IEEE Transactions on Biomedical Engineering},
+   volume = {54},
+   number = {11},
+   pages = {1940-1950},
+   ISSN = {1558-2531},
+   DOI = {10.1109/TBME.2007.901024},
+   year = {2007},
+   type = {Journal Article}
+}
+
+@online{SciPy:2021,
+   author = {SciPy-Developers},
+   title = {SciPy Spatial Transformations},
+   url = {https://docs.scipy.org/doc/scipy/reference/spatial.transform.html},
+   year = {2021},
+   urldate = {2021-09-13}
+}
+
+@techreport{Diebel:2006,
+   author = {Diebel, James},
+   title = {Representing Attitude : Euler Angles , Unit Quaternions , and Rotation Vectors},
+   institution = {Stanford University},
+   year = {2006},
+   type = {Report}
+}
+
+@online{Visual3D:2021,
+   author = {CMotion},
+   title = {CMotion Visual3D},
+   url = {https://www.c-motion.com/#visual3d},
+   year = {2021},
+   urldate = {2021-09-13}
+}
+
+@online{AnyBody:2020,
+   author = {AnyBodyTech},
+   title = {AnyBody Technology},
+   url = {https://www.anybodytech.com},
+   year = {2020},
+   urldate = {2021-09-13}
+}
+
+@online{Wikipedia:2021,
+   author = {Wikipedia},
+   title = {Euler angles},
+   url = {https://en.wikipedia.org/wiki/Euler_angles},
+   year = {2021},
+   urldate = {2021-09-13}
+}
+
+@book{Zatsiorsky:1998,
+   author = {Zatsiorsky, Vladimir M.},
+   title = {Kinematics of Human Motion},
+   publisher = {Human Kinetics},
+   ISBN = {0-88011-676-5},
+   year = {1998},
+   type = {Book}
+}
