+++ joos/paper.md
@@ -32,18 +32,18 @@ bibliography: paper.bib
 
 # Summary
 
-Kinematic analysis studies characterize human motion in healthy, pathologic, and rehabilitated subjects. These studies provide rich datasets which enable clinicians and researchers to understand disease progression, and the effects of surgical intervention and physical therapy. A variety of techniques – ranging from optical skin marker tracking to biplane fluoroscopy – are utilized to collect these datasets. The quantification and physical interpretation of these motion capture datasets (i.e. kinematic analysis) are predominantly conducted using Euler/Cardan angles, whose output - time series of 3 angles - is presented in peer-reviewed manuscripts in 2D graphs [@Ludewig:2009]. This rudimentary presentation, however, obscures the physical meaning behind Euler/Cardan analysis. As other authors have noted, the visualization of these angles enables meaningful discussion between biomechanics researchers and clinicians [@Baker:2011].  A ubiquitous open-source method to concurrently visualize recorded human motion and the results of kinematic analysis is presently lacking in the biomechanics community. ‘Kinematics-vis’ is a JavaScript web application that visualizes kinematic analysis output resulting from motion capture studies.
+Kinematic analysis studies characterize human motion in healthy, pathologic, and rehabilitated subjects. These studies provide rich datasets which enable clinicians and researchers to understand disease progression, and the effects of surgical intervention and physical therapy. A variety of techniques – ranging from optical skin marker tracking to biplane fluoroscopy – are utilized to collect these datasets. The quantification and physical interpretation of these motion capture datasets (i.e. kinematic analysis) are predominantly conducted using Euler/Cardan angles [@Wikipedia:2021; @Diebel:2006; @Zatsiorsky:1998], whose output – time series of 3 angles – is presented in peer-reviewed manuscripts in 2D graphs [@Ludewig:2009]. This rudimentary presentation, however, obscures the physical meaning behind Euler/Cardan analysis. As other authors have noted, the visualization of these angles enables meaningful discussion between biomechanics researchers and clinicians [@Baker:2011].  A ubiquitous open-source method to concurrently visualize recorded human motion and the results of kinematic analysis is presently lacking in the biomechanics community. ‘Kinematics-vis’ is a JavaScript web application that visualizes kinematic analysis output resulting from motion capture studies.
 
 
 # Statement of need
 
-3D rotations are utilized to quantify human motion and in kinematic analysis. However, because 3D rotations belong to the SO(3) mathematical group, they can be difficult to understand – especially for biomechanics researchers and clinicians with a limited mathematical background. Euler/Cardan angles are typically utilized to infuse physical meaning into kinematic analysis. However, they are described abstractly and attempts towards visualizing their physical underpinning is rarely made [@Baker:2011]. Furthermore, 12 Euler/Cardan angle sequences exist for quantifying 3D orientations, and which sequence is most appropriate for quantifying human motion is debatable and varies by study [@Krishnan:2019]. It has previously been shown that incorrect usage of Euler/Cardan sequences can significantly misrepresent human motion [@Aliaj:2021]. Furthermore, studies rarely present multiple Euler/Cardan sequences in publications – which clouds data interpretation and limits the ability to compare results between laboratory groups and studies.
[…]
