+++ js/EulerScene_SphereAreaVis.js
@@ -1,8 +1,8 @@
 import {EulerScene} from "./EulerScene.js";
-import {Group, Matrix4, Mesh, Quaternion, Vector3} from "./vendor/three.js/build/three.module.js";
-import {LineMaterial} from "./vendor/three.js/examples/jsm/lines/LineMaterial.js";
-import {LineGeometry} from "./vendor/three.js/examples/jsm/lines/LineGeometry.js";
-import {Line2} from "./vendor/three.js/examples/jsm/lines/Line2.js";
+import {Group, Matrix4, Mesh, Quaternion, Vector3} from "three";
+import {LineMaterial} from "three/examples/jsm/lines/LineMaterial.js";
+import {LineGeometry} from "three/examples/jsm/lines/LineGeometry.js";
+import {Line2} from "three/examples/jsm/lines/Line2.js";
 
 EulerScene.prototype.prepare_scene_for_area_vis = function(quatFnc) {
     const recursiveEnable = child => child.layers.enable(this.areaVisLayer);
