### 2021-09-21 klevis-a/kinematics-vis#4 @klevis-a
@nnadeau I actually got around to using standard packaging (yarn and webpack). See https://github.com/klevis-a/kinematics-vis/pull/20.
