### 2021-09-23 5275346
Removed downloadData scripts and switched to Gulp for downloading data.
