### 2021-09-23 4d86acc
Added another entrypoint to the application (single.html). Common JS files are shared between main.html and single.html.
