### 2021-11-02 #3490 @adi3
@destogl - I see you're almost done with your checklist. Could you please plan to finish it sometime this week so we can move ahead with publishing this paper? Greatly appreciate your help!
