### 2021-12-11 #3490 @destogl
Dear all,

Thank you for your patience and for the opportunity for his review. I am deeply sorry that my review took so long, I will work to change this in the future and check my notifications regularly.

@adi3 my review is finished.

@klevis-a and @heath-henninger congratulation for the great work! I am fully supporting acceptance of its publication.
The improvements @klevis-a did after my review are great and now I as non-biomechanic person can understand how the software works.

A one closing note: @klevis-a  I think you can close klevis-a/kinematics-vis#8 it does not seem to be relevant anymore.

I wish you all much success in the future and to continue producing great open-source software :)
