### 2021-12-11 9f10823
Added link to MANUAL_TESTING.md from README.md.
