### 2021-12-11 #3490 @heath-henninger
@destogl I echo @klevis-a's comments, thank you for the kind words and review!  We hope this platform will find wide utilization in biomechanics laboratories and courses to help learning and reinforcing kinematics concepts more intuitively.
