### 2021-12-13 #3490 @adi3
Hurray! @destogl thank you so much for your thorough review! We appreciate all the effort you put into it.

@klevis-a @heath-henninger - I have just one more formality required from you. Can you please issue a new tagged release of the software and archive it on Zenodo, figshare or another archiving platform of your choice? Then report the release's version number and the archive's DOI on this thread. Thank you, and congratulations!
