### 2021-12-13 #3490 @heath-henninger
@adi3, we will likely use Zenodo as that is where our other open access kinematic data sets that drive Kinematics-vis live currently.  Will that Zenodo DOI then be listed in the published manuscript?  Will it move forward from DRAFT status after that link is published?  Sorry for the questions...this is the first time we have published using the JoOS mechanism and it's quite different than the common print publications we deal with!
