### 2021-12-14 #3490 @klevis-a
Thank you, @adi3!

Just released version [0.2.0](https://github.com/klevis-a/kinematics-vis/releases/tag/0.2.0). This is archived on Zenodo with a DOI of [10.5281/zenodo.5781772](https://zenodo.org/record/5781772).
