### 2021-12-21 #3490 @adi3
@heath-henninger yes, the DOI will be published in the manuscript. I believe I have everything from your side now to push this to publication.
