### 2021-12-21 #3490 @klevis-a
Thank you @arfon and thank you @adi3! Working through the review process for JOSS has been cool. I very much enjoyed the transparency and automation. I am thankful for the acceptance and for all the hard work everyone has put into the journal.

@destogl, @nnadeau thank you for taking the time to review the manuscript and software, and for your feedback!
